test: Add integration tests for logger functionality with NATS
- Introduced a new integration test file for the logger, focusing on testing interactions with a real NATS server using testcontainers.
- Implemented multiple tests to validate the logger's message writing, rotation, and compression functionalities.
- Enhanced test coverage for integration scenarios, ensuring reliable setup and teardown of test environments with NATS.
- Updated existing tests for clarity and improved naming conventions.
