+++ cmd/logger/main_test.go
@@ -12,8 +12,8 @@ import (
 	"github.com/saviobatista/sbs-logger/internal/types"
 )
 
-// TestEnvironmentVariables tests environment variable handling
-func TestEnvironmentVariables(t *testing.T) {
+// TestParseEnvironment tests environment variable parsing
+func TestParseEnvironment(t *testing.T) {
 	// Save original environment
 	originalOutputDir := os.Getenv("OUTPUT_DIR")
 	originalNATSURL := os.Getenv("NATS_URL")
@@ -155,90 +155,6 @@ func TestLogger_Start(t *testing.T) {
 			}
 		})
 	}
-}
-
-// TestLogger_RotationTimer tests the rotation timer with context cancellation
-func TestLogger_RotationTimer(t *testing.T) {
-	tempDir, err := os.MkdirTemp("", "logger-test")
-	if err != nil {
-		t.Fatalf("Failed to create temp dir: %v", err)
-	}
-	defer func() { _ = os.RemoveAll(tempDir) }()
-
-	logger := NewLogger(tempDir)
-
-	// Initialize the logger first
-	err = logger.rotateFile()
-	if err != nil {
-		t.Fatalf("Failed to initialize logger: %v", err)
-	}
-	defer func() {
-		if logger.GetCurrentFile() != nil {
-			_ = logger.GetCurrentFile().Close()
-		}
-	}()
-
-	// Test context cancellation
-	ctx, cancel := context.WithCancel(context.Background())
-
-	// Start rotation timer in goroutine
-	done := make(chan bool)
-	go func() {
-		logger.rotationTimer(ctx)
-		done <- true
-	}()
-
-	// Cancel context after a short delay
-	go func() {
-		time.Sleep(10 * time.Millisecond)
-		cancel()
-	}()
-
-	// Wait for function to return due to context cancellation
-	select {
-	case <-done:
-		// Function returned as expected
-	case <-time.After(1 * time.Second):
-		t.Error("rotationTimer did not return when context was cancelled")
-	}
-}
-
-// TestLogger_RotationTimer_RotationTrigger tests rotation trigger functionality
-func TestLogger_RotationTimer_RotationTrigger(t *testing.T) {
-	tempDir, err := os.MkdirTemp("", "logger-test")
-	if err != nil {
-		t.Fatalf("Failed to create temp dir: %v", err)
-	}
-	defer func() { _ = os.RemoveAll(tempDir) }()
-
-	logger := NewLogger(tempDir)
-
-	// Initialize the logger first
-	err = logger.rotateFile()
-	if err != nil {
-		t.Fatalf("Failed to initialize logger: %v", err)
-	}
-	defer func() {
-		if logger.GetCurrentFile() != nil {
-			logger.GetCurrentFile().Close()
-		}
-	}()
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	// Start rotation timer
-	go logger.rotationTimer(ctx)
-
-	// Trigger rotation
-	logger.rotationChan <- struct{}{}
-
-	// Give time for rotation to process
-	time.Sleep(50 * time.Millisecond)
-
-	// The rotation should have been processed (this tests the rotation channel case)
-	// We can't easily verify the exact behavior without more complex mocking,
-	// but the coverage will be improved by exercising this code path
 }
 
 // TestLogger_RotateFile tests file rotation with error scenarios
@@ -397,105 +313,6 @@ func TestLogger_WriteMessage(t *testing.T) {
 	}
 }
 
-// TestLogger_RotateAndCompress tests comprehensive rotation and compression scenarios
-func TestLogger_RotateAndCompress(t *testing.T) {
-	tests := []struct {
-		name        string
-		setupLogger func() (*Logger, func())
-		expectError bool
-	}{
-		{
-			name: "successful rotation and compression",
-			setupLogger: func() (*Logger, func()) {
-				tempDir, err := os.MkdirTemp("", "logger-test")
-				if err != nil {
-					panic(err)
-				}
-				logger := NewLogger(tempDir)
-
-				// Set up initial state with a previous date
-				yesterday := time.Now().UTC().AddDate(0, 0, -1).Format("2006-01-02")
-				logger.SetCurrentDateForTesting(yesterday)
-
-				err = logger.rotateFile()
-				if err != nil {
-					panic(err)
-				}
-
-				// Write some content
-				_, _ = logger.GetCurrentFile().WriteString("test content\n")
-
-				return logger, func() { os.RemoveAll(tempDir) }
-			},
-			expectError: false,
-		},
-		{
-			name: "rotation with no current file",
-			setupLogger: func() (*Logger, func()) {
-				tempDir, err := os.MkdirTemp("", "logger-test")
-				if err != nil {
-					panic(err)
-				}
-				logger := NewLogger(tempDir)
-				// Don't initialize current file
-				logger.SetCurrentDateForTesting(time.Now().UTC().AddDate(0, 0, -1).Format("2006-01-02"))
-				return logger, func() { os.RemoveAll(tempDir) }
-			},
-			expectError: false,
-		},
-		{
-			name: "rotation with no current date (no compression)",
-			setupLogger: func() (*Logger, func()) {
-				tempDir, err := os.MkdirTemp("", "logger-test")
-				if err != nil {
-					panic(err)
-				}
-				logger := NewLogger(tempDir)
-
-				err = logger.rotateFile()
-				if err != nil {
-					panic(err)
-				}
-
-				return logger, func() { os.RemoveAll(tempDir) }
-			},
-			expectError: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			logger, cleanup := tt.setupLogger()
-			defer cleanup()
-			defer func() {
-				if logger.GetCurrentFile() != nil {
-					logger.GetCurrentFile().Close()
-				}
-			}()
-
-			err := logger.rotateAndCompress()
-
-			if tt.expectError && err == nil {
-				t.Error("Expected error, got none")
-			}
-			if !tt.expectError && err != nil {
-				t.Errorf("Expected no error, got: %v", err)
-			}
-
-			if !tt.expectError {
-				// Verify new file was created with today's date
-				expectedDate := time.Now().UTC().Format("2006-01-02")
-				if logger.GetCurrentDate() != expectedDate {
-					t.Errorf("Expected current date %s, got %s", expectedDate, logger.GetCurrentDate())
-				}
-				if logger.GetCurrentFile() == nil {
-					t.Error("Expected current file to be set after rotation")
-				}
-			}
-		})
-	}
-}
-
 // TestCompressFile tests comprehensive compression scenarios
 func TestCompressFile(t *testing.T) {
 	tests := []struct {
@@ -524,28 +341,6 @@ func TestCompressFile(t *testing.T) {
 			name: "compress non-existent file",
 			setupFile: func() (string, func()) {
 				return "/tmp/non-existent-file.log", func() {}
-			},
-			expectError:   true,
-			errorContains: "failed to read file",
-		},
-		{
-			name: "compress file in read-only directory",
-			setupFile: func() (string, func()) {
-				tempDir, err := os.MkdirTemp("", "logger-test")
-				if err != nil {
-					panic(err)
-				}
-				testFile := filepath.Join(tempDir, "test.log")
-				err = os.WriteFile(testFile, []byte("test content\n"), 0600)
-				if err != nil {
-					panic(err)
-				}
-				// Make directory read-only to prevent reading the file
-				_ = os.Chmod(tempDir, 0400)
-				return testFile, func() {
-					_ = os.Chmod(tempDir, 0600)
-					_ = os.RemoveAll(tempDir)
-				}
 			},
 			expectError:   true,
 			errorContains: "failed to read file",
@@ -646,70 +441,6 @@ func TestLogger_ErrorScenarios(t *testing.T) {
 			t.Errorf("Expected 'failed to write message' error, got: %v", err)
 		}
 	})
-}
-
-// TestLogger_Integration tests the full logger workflow
-func TestLogger_Integration(t *testing.T) {
-	// Create temporary directory
-	tempDir, err := os.MkdirTemp("", "logger-test")
-	if err != nil {
-		t.Fatalf("Failed to create temp dir: %v", err)
-	}
-	defer func() { _ = os.RemoveAll(tempDir) }()
-
-	// Setup logger
-	logger := NewLogger(tempDir)
-	ctx, cancel := context.WithTimeout(context.Background(), 200*time.Millisecond)
-	defer cancel()
-
-	// Start logger
-	go logger.Start(ctx)
-
-	// Give it a moment to initialize
-	time.Sleep(50 * time.Millisecond)
-
-	// Write multiple messages
-	messages := []*types.SBSMessage{
-		{
-			Raw:       "MSG,1,1,1,ABC123,1,2021-01-01,00:00:00.000,2021-01-01,00:00:00.000,,,,,,,,,,\n",
-			Timestamp: time.Now(),
-			Source:    "test-source-1",
-		},
-		{
-			Raw:       "MSG,3,1,1,DEF456,1,2021-01-01,00:00:01.000,2021-01-01,00:00:01.000,TEST456,11000,500,200,41.7128,-75.0060,0,0,0,0\n",
-			Timestamp: time.Now(),
-			Source:    "test-source-2",
-		},
-	}
-
-	for _, msg := range messages {
-		err = logger.WriteMessage(msg)
-		if err != nil {
-			t.Errorf("Failed to write message: %v", err)
-		}
-	}
-
-	// Wait for context to complete
-	<-ctx.Done()
-
-	// Verify log file was created and contains messages
-	expectedDate := time.Now().UTC().Format("2006-01-02")
-	logPath := filepath.Join(tempDir, fmt.Sprintf("sbs_%s.log", expectedDate))
-
-	content, err := os.ReadFile(logPath)
-	if err != nil {
-		t.Fatalf("Failed to read log file: %v", err)
-	}
-
-	expectedContent := messages[0].Raw + messages[1].Raw
-	if string(content) != expectedContent {
-		t.Errorf("Expected log content %q, got %q", expectedContent, string(content))
-	}
-
-	// Cleanup
-	if logger.GetCurrentFile() != nil {
-		logger.GetCurrentFile().Close()
-	}
 }
 
 // Helper functions
